Ensure displayId isn't translated
String.format() will also translate some strings. When the phone's
language is set to something that doesn't use the Western Arabic numeral
system, these numbers are translated as well, which means that the
display device config files cannot be found, as they are described by
their display id. This cl ensures the Locale is the same as the root locale when calling String.format() for the ddc path, meaning they won't get translated.

Bug: 241871531
Test: adb shell dumpsys display | grep DisplayDeviceConfig
Change-Id: I10b022c04c743f2eb287426a8a9026cc0c75c276

Former-commit-id: 7c94b7a9ef3b37e31d320e351caf0fe058e20bdf