Set rotation layers as opaque initially
The layers should cover the window behind during rotation animation.
So it can be opaque initially to avoid blended composition for one
intermediate frame.

After the animation starts and changes the alpha < 1 to the rotation
layer for fading animation, it can still switch from opaque composition
to blended composition automatically.

Bug: 240568519
Test: atest ChangeAppRotationTest

Change-Id: Ic4173d33c15c35e252315f9cf53127c759e56ece

Former-commit-id: 73d7468fe62fea979d12b1f33c5f807e62f1a0e0