Always add cookie for each activity when fill TaskInfo.
The LaunchCookie is a token for activity starter to find out where the
activity was start, so it should be allowed to add in TaskInfo even
when the activity is finishing.

Bug: 241070422
Test: Launch an activity which will request finish when receive back
key from Launcher, e.g. calendar widget. Verify
QuickstepTransitionManager can find the LauncherView so it can play
closing to home animation.
Test: atest ActivityRecordTests#testTransferLaunchCookieWhenFinishing

Change-Id: I0351eb76e9a97b718b490d6151b5434f8ef8a42c

Former-commit-id: 9550f1c9496c1550c09890a4b7dc8cb5976ae4d6