Removes the redundant instance for singleInstance
The activity with launchMode "singleInstance" was able to be created
in multiple instances. It contradicts with developers' documentation.
The conditions for searching the reusable tasks always ignore the
start-for-result results in singleInstance instances being created.

The CL finishes existing instance when a new instance of
singleInstance activity launches, in order to keep only one instance
in the system. Some compatibility issues might be introduced if app
incorrectly depend on singleInstance activities to be multiple
instances.

Bug: 122967919
Test: atest IntentTests
Change-Id: I6345985a0ee5bd47d100d529bc3b50e8ae4d2bab

Former-commit-id: 9a0b544d9303516e18405f7f91c6444f468c9a14