Fix bubbles hiding when unlocked with showOnLockscreen Activity
KeyguardStateController#isOccluded was being used to determine if
the device was showing the screen saver / dreams. However, this
method returns true even if the device is in unlocked state.

This can happen when an activity is showing that has showOnLockscreen
set to true even while the device is unlocked.

This CL uses DreamManager to check if the device is dreaming instead
of KeyguardStateController#isOccluded which only returns true if
the device is actually showing screen saver / dreams.

Bug: 240510360
Test: atest BubblesTest
Test: manual - open calculator and verify bubbles remain visible
             - start dream and verify bubbles are hidden
Change-Id: I6997c17d26c2f8ccbe80a5920f9348b3ce53d754

Former-commit-id: ddf81fc112ff6ee25e9205ebea3dbeae93fbfcb1