Add a ktfmt hook for some SystemUI files
This CL enables a ktfmt hook to check that Kotlin files are properly
formatted.

The set of files or directories that should be included is specified in
a simple ktfmt_includes.txt file. That way, existing files that are not
properly formatted yet won't be checked and uploaders won't be bothered
by this check. In the future, we will check all files created after some
baseline.

At the moment, only Kotlin files that I know are already formatted with
ktfmt were included in that list.

Test: Manual
Bug: 235461679
Change-Id: Idaed15c39f6af413c305f4c74f142a2895b5a753

Former-commit-id: e261526d2c9113382bee2d9d1cb8663f0e360585