Fix wrong path is returned when 2 files have the same name
Before the fix, the cache is reused when 2 files have the same name.
Now the full file path is included when calculating the cache key.
This will reduce the chance of key collision. We also check if the
paths match to ensure 2 different files are never mapped to the same
cache entry.

Bug: 242639590
Test: atest PackageParserTest
Change-Id: I4d240407f6fd6de9a64f9ea46339d73202d40927

Former-commit-id: eb832a8a336274260be8490d8aa3ad8592cd87f2