Provide a way to keep dreaming when undocking.
This change makes it possible to keep a device dreaming when it is
removed from its dock. But there should be no change to existing
functionality unless config_keepDreamingWhenUndocking is true.

Bug: 239090297
Test: Enable screen saver and set it to start when device is docked.
Dock device and wait for screen saver to appear. Remove device from dock
and verify that the screen saver ends (the existing behavior without
this change).

Change-Id: I345eb1d17ae88d893789dfcc160b4a9a664ba57f

Former-commit-id: bef02552736b17249b15f506bd4b8f7caf17ce12