Enables flag for modern architecture in KBAV.
The KeyguardBottomAreaView (KBAV) has been refactored to a new, modern
architecture. This CL flips the flag to turn that on. The plan is to let
it soak for a while with Droidfooders and fix any bugs they find. Once
we are happy, we will delete this flag and all the now-dead code
protected behind it.

Bug: 235403546
Test: the modern architecture is used even when I turn off the teamfood
flag in the flag flipping app.

Change-Id: I36b4c6d28e658d7dc55ec03f448a8ca728855f94

Former-commit-id: db54e85268da34a24e585f95396d4820b2d30242