Moving autobrightness display configs to a different file
Currently, all the display configs are controlled from
display_device_config.xsd file. This has led to a lot of configs being
controlled from this single source, making it error prone and less
readable. This is to decouple that, and move the autobrightness related
display configs to a separate file. We choose to move autobrightness
configs because we soon plan to add more configs to this.

Bug: 235345267
Test: atest DisplayDeviceConfigTest#testConfigValues
Change-Id: Ia3ba1065663f478ffcb6f4ea5951958352679f76

Former-commit-id: c45cb706ff191f9b70550e9b99036279c390b747