[SB Refactor] Define a new status bar wifi view that can be swapped in/out with the old view. Define a skeleton ViewBinder for the new view.
This CL sets us up so that, if we flip a flag, we can start rendering
the status bar wifi icon using the new pipeline instead of the old
pipeline. Right now, it's always just a wifi level 2 icon, but in future
CLs it will get hooked up to the real data sources and update
accordingly.

To test:
adb shell cmd statusbar flag 604 on // Turn on the new backend
adb shell cmd statusbar flag 605 on // Turn on the new frontend

Then, you should see a cyan wifi icon in the status bar. (Cyan for
debugging purposes, so we know it's a new pipeline icon.)

Most of this CL is just piping values to the right places, since the
status bar icon manager isn't fully dagger-ified yet.

Test: manual: verified that flipping flag 605 on/off toggles the wifi
icon between the old pipeline and new pipeline.
Test: atest SystemUITests
Bug: 238425913

Change-Id: I60cbc17e6677aad51c993402d7497bf670ab9caa

Former-commit-id: 54db672485131ef001301517aedddbf823e709fc