Migrate back animation to shell transition
When back animation finished, it will invoke the real back callback and
cause a new transition started.

In this CL, we introduce the back transition handler to consume the
incoming transition request and takeover the whole transition if the
transition info contians same departing window token.

This also seperated the behaviors of enabled/disabled shell
transition.

Bug: 238475694
Test: Enabled shell transition, atest BackNavigationControllerTests
      BackAnimationControllerTest
Change-Id: I57e7c89ce6cb7a99ab3af403704b9dd948f26151

Former-commit-id: 5c85e5b6c09b2a60b14d3c74ee9bd01c7d338f35