Fix blank screen when launching controls activity
It was possible to attempt to launch a non showWhenLocked activity
over the keyguard by passing the wrong value into the
activityStarter. In addition, the existing code would trampoline
activities which produced a bad opening animation.

Directly launch the correct activity, and also set the
showOverLockscreenWhenLocked flag to the correct value to ensure the
animation runs properly.

Fixes: 235769893
Test: atest DeviceControlsTileTest
Test: manual, follow steps in bug to repro
Change-Id: I64ebd82b47547beecaee15d570a14dc20fe663ef

Former-commit-id: 987035d3653a19dadb11a436e6c2d44c28abe0fb