Migrate GlobalActionsDialogLite to predictive back
As part of the effort to rollout Predictive Back, the old
"just in time" back handlers need to be migrated over to the new
"ahead of time" api (before any animation/feedback can be added).

This is the first change in this API migration, with more to come.

This also adds tests for GlobalActionsDialogLite.

Bug: 238358031
Test: atest GlobalActionsDialogLiteTest
Change-Id: Id8ee4ec97e34e6de63659d30784384e7c9ef9590

Former-commit-id: 1dde50349e9a267ce2f3a3700f6d7acf31e5544c