Add some fakes to SystemUI-test-utils
This CL adds fake implementations for:
 - FgsManagerController
 - UserTracker
 - SecurityController
 - UserInfoController

It also adds a wrapper around a mocked UserSwitcherController to make it
easier to mock in tests while still handling callbacks.

Those fakes are going to be used to test the new implementation of the
QSFooterActions, which are going to be refactored using the modern
Android architecture. See ag/19678215 to see how they are used.

Bug: 242040009
Test: atest FgsManagerControllerTest
Test: atest FooterActionsViewModelTest
Test: atest FooterActionsInteractorTest
Change-Id: I0a3ef36de81c1bd48f33d4d234a4ee903bf40fc6

Former-commit-id: e496b54c2129615a9c7513ee5007d6fa49b31f75