Fix shade collapse on keyguard with bypass enabled
Only with face unlock + bypass enabled, when dragging down the shade
over keyguard, the end motion event was forcing the shade to collapse
again, due to an incorrect check on the keyguard state. Use the
KeyguardStateController.isShowing() method instead, which is a more
accurate check on keyguard state.

Fixes: 242813247
Test: All combinations of face unlock enabled/disabled, notifications
visible on lockscreen on/off, notification redacted on/off, face
bypass on/off

Change-Id: I99c6ced66f3269dbf4e6df1d078d7c793492c157

Former-commit-id: e11a139facc0fca468cbb971aa4a5965fe9040a4