Fix bug with clearing bubble to expand
When BubbleController receives multiple status bar state updates, it can
clear the bubble to be expanded information.
This can happen when user receives a bubble notification on lockscreen
and taps on it. In this case we set the bubble to be expanded field on
the controller as the device is locked and we can't expand the bubble
yet.
Phone unlock flow triggers multiple statusbar updates to the
BubbleController which can clear the bubble to be expanded field.
Make sure field is only cleared when bubble actually is going to be
expanded.

Bug: 240226362
Test: atest BubblesTest
Test: have bubbles stacked, pattern lock screen, lock device and receive
an incoming notification to one of the bubbles, click on the
notification to unlock device. Observe bubbles is expanded after
unlocking device.

Change-Id: I61f1d6a1fb673a89fa9b7a1d6b6253874d0c5778

Former-commit-id: b72b72ac13a5ac3f2d4ea8184eb0b2f2460c34ca