Cleanup dream complication setting to reflect single toggle users see in Settings.
The original implementation was designed to allow users to individually
enable/disable certain complications. However, we have since decided to
simplify the complication controls around a single master toggle. This
fixes an issue where when new complications are added, we will no longer
switch the toggle to off in Settings.

Bug: 241469874
Test: atest ComplicationTypesUpdaterTest
Test: atest DreamBackendTest
Test: manually on device
Change-Id: I1bbfbbe2abd6398b74bcd0b9ba601583a1403bef

Former-commit-id: 07cfa206f46ad6d7cb18925563e1fdf6f335f1ce