Add media entry Dream complication.
This implements using the entry chip to show/hide UMO, not yet the
timeout behavior. Also, the chip visual isn't updated when it's clicked.

The layout weight for the UMO complication is updated to ensure it's
placed below the Smartspace, not above it.

Also fixed a javadoc broken link in ComplicationUtils.java.

Bug: 242613318
Test: manual
Test: atest MediaDreamSentinelTest
Change-Id: Ic1eb7585007493be509ccce50c68e2d5a91583fe

Former-commit-id: ea8cd10083ac4d257b4f9b002b50a3789e868848