Add runtime unmodifiable check for PackageState
Introduces test that enforces that all members and their type
parameter children types that return collections/maps return non-null,
non-empty, unmodifiable objects when a test instance is constructed
and filled with data.

Makes the PackageImpl collections unmodifiable, but ignores the inner
objects with a constants list in the test. To be removed once the
inner object infrastructure is finished and can be marked immutable.

Bug: 235462722

Test: atest PackageManagerServiceUnitTests

Change-Id: I9759b7dd5c945e65498e22fcf408d53931dc502f

Former-commit-id: fad1ef9961087734b7864c3ff074eb28b27db5f3