HDMICEC: Remove deferred OneTouchPlayAction
This fix a critical bug of playback devices sending an <Active Source> message when they are not the source selected by users, taking over TVs screen. This situation occurs when users are changing input sources quickly, triggering a playback device to wake up from the standby state, starting CEC initialization. When the initialization is complete, the device tries to start an OneTouchPlayAction, which will trigger the <Active Source> message to be broadcasted.

The change includes setting mAddressAllocated to true after the onInitializeCecComplete call as opposed to before. This way, a possible created OneTouchPlayAction is not started right away, but deferred. This action is cancelled when preprocessing the buffer messages.

Bug: 235434112
Test: make && atest HdmiCecLocalDevicePlaybackTest
Change-Id: Ic189ba5ab72b8e8b4f9d957f3367b453e82df8fd

Former-commit-id: d4927bbaa28adefb2b7fa5d1a9d82552d2898f37