Set the Low Power Standby enabled setting when unassigned
SettingsSnapshot metrics allow collecting data on the current value of
Settings, but if the Setting has never been written to, a Setting will
be logged as <notassigned>.

We want to track the standby mode users are using. We can do that by
collecting data on the LOW_POWER_STANDBY_ENABLED setting, but since
some devices may have LPS enabled by default while others don't, we
can't tell from a <notassigned> value whether LPS is enabled or not on a
device.

With this change, when the Setting hasn't been written yet, then the
Setting is updated to the actual value. Any <notassigned> value then
means that LPS is not supported, as otherwise the value would've been
assigned.

Test: manual - check whether setting gets written
Bug: 234771100
Change-Id: I62ebe69d01bce99c78ed21326538e38fa6f21428

Former-commit-id: 6dbc4a144f4441421c2376b06edcbc4aa69a0e9f