Enforce TaskFragmentOrganizer can only manipulate organized TF
When TaskFragmentOrganizer apply a WCT using TaskFragment token, we
should also enforce that the TaskFragment is organized by the caller,
because when an app is embedded by another app, the embedee may be
able to get the fragment token from
ActivityClientRecord#mInitialTaskFragmentToken.

Also remove check for TYPE_REORDER because it is currently only
supported for Task, and we need to re-evaluate cross-process embedding
overlay if we want to allow reorder.

Bug: 243741794
Bug: 243330085
Test: atest WmTests:TaskFragmentOrganizerControllerTest
Change-Id: I92afbddd7086c606e55cdf2982d9a241adde8db9

Former-commit-id: 345baf632ffd6653b665ff699ca354952c4cc5fb