Factor out broadcast "skip" policy.
Upcoming work will look at making significant changes to the
BroadcastQueue internals, so this change factors out the current
implementation to aid A/B testing.

This change migrates the existing "skip" logic into a dedicated
location where it can be shared.  This change is designed to be a
no-op refactoring with no actual behavior changes.

Bug: 243656033
Test: atest CtsContentTestCases:BroadcastReceiverTest
Change-Id: I8bc3e81286ba0b548533e60d4a47ceb43714f5b1

Former-commit-id: 676adec23816a116a0e27549f56a44d967621b5d