Disable screenshot tests in master (1/2)
This CL disables SystemUI screenshot tests in master given that the
golden images are generated from the tm-qpr-dev branch which can be
different from the master branch in term of how text, etc are rendered.
Given that all of SystemUI is working in tm-qpr-dev, we can have the
screenshot tests only there.

Bug: 241721843
Test: Manual
Change-Id: Ie7afe9fe7e05742b9ac485371cfdf97dbe511146

Former-commit-id: 3c337ad32506fe242be95a7963c6ec4cea300fce