First pass at setting up a mobile connectivity pipeline
This CL is pretty WIP, but contains a rough overall structure for
monitoring mobile subscriptions, and info for each subscription.

Naming is still TBD. The general structure is this:

[Repo]
MobileSubscriptionRepo monitors the list of subscriptions known to the
device via SubscriptionManager, and can vend out a listener for a given
subscription ID that can be individually tracked

UserSetupRepo just tracks the isUserSetup bit

[Domain]
MobileInteractor defines the Interactor for the overall mobile feature,
and can vend out MobileConnectionInteractors which track the individual
subscriptionID from the repo

[UI]
Currently this is the least fleshed out, but there is a similar pair of
MobileConnectivityViewModel and MobileIconViewModel to track the same as
above -- the overall state and the states for each subId.

Test: adb shell settings put global systemui/tag/SbConnectivity verbose
Test: adb logcat | grep SbConnectivity
Test: should output: SbConnectivity: Output: subId(1) level: 1
Bug: 240492102
Change-Id: I22d7e9e5b449057a92071ac00bc823460d5f58a6

Former-commit-id: 97592f60b974051d3490acd7078cfc4cb1fadbaa