Fix key frames in headers animation
When setting values for `sizePercent` (or `percentX/Y`) in a KeyPosition,
the value is the relative growth from 0 (beginning of animation) to 1
(end of animation).

Therefore, in order to not have apparent size change
(and movement) before and after the fade in the expansion, we need to
set it to 0 for the `KeyPosition` before 0.5 and 1 for the one after.

Test: manual, date and icons don't move horizontally on expansion
Fixes: 243260726
Change-Id: I30fcb6f4f0a8f9cae98aba22c0400f9c5a8513f1

Former-commit-id: 33d3c266616ffdf0f56159d5cddd95ba5137ae9e