Fix letterbox for SCREEN_ORIENTATION_BEHIND
At the moment, an activity with "behind" doesn’t go in
letterbox as expected when it’s at the top of a portrait-only
one which was letterboxed in landscape device orientation.

An activity with "behind" orientation will be letterboxed if it is
the activity that's immediately beneath it

Test: Launch an activity using SCREEN_ORIENTATION_BEHIND on top
of a fixed orientation activity and check it's letterboxed. Also run
`atest ActivityRecordTests`

Fix: 244309481

Change-Id: I221e4c2a778a4e2ea11795e82366bf5bb08a5507

Former-commit-id: dd0f6850606c85b6914e1f946efd32a132bae599