Changes user start to enforce profiles are assigned to the parent display.
Previously, when ActivityManager.startProfile() was called, it would
trigger starting the profile on DEFAULT_DISPLAY, which would fail if
called by a user running on background in a secondary display (as the
server side part would require the caller to pass the parent's
display id).

This change fixes that behavior by using a special display id
(Display.PARENT_DISPLAY), which is only used when the device supports
multiple users on multiple displays.

Test: atest FrameworksMockingServicesTests:com.android.server.pm.UserManagerServiceTest UserManagerInternalTest UserControllerTest

Bug: 239982558
Bug: 244452695

Change-Id: I92aa2df97b3a6cf71d9386e67111ff8da6eedf56

Former-commit-id: f7e65e58821c0f6f71b28c74e673d79d53e2d429