When asleep, dismiss inattentive sleep warning, don't reschedule
In the PowerManagerService power state update method the inattentive
sleep warning was dismissed when the device wasn't awake.
This incorrectly also happened when the device was dreaming, dismissing
the warning only to show it immediately again, causing some flicker.
Also, the message triggering the warning to be shown and the device to
go to sleep was still rescheduled when the device was already asleep,
which is unnecessary.

Bug: 243491344
Test: atest PowerManagerServiceTest#testInattentiveSleep_warningStaysWhenDreaming
Test: atest PowerManagerServiceTest#testInattentiveSleep_warningNotShownWhenSleeping
Change-Id: Iaddf84b107a231467653dd98b46839579444a56a

Former-commit-id: 8b5c186db5bd49827b3a89a827d83e8bce07237b