Do not create extend surface for transfer starting window change.
There could transfer starting window when launch a trampoline activity,
and the trampoline activity could be removed before it add a window.
So a possible sequence would be:

1. Activity A start a trampoline activity B, and transfer the starting
window to B.
2. B finish itself before add window, so the starting window transfer
back to A.

And there will create a close transition when finish B, but actually
this activity never draw anything on it. So if there attemp to create
an edge extend surface for it, that thread would be stuck in
Surface.unlockAndPost because nothing will draw.
Note: Not sure if above reasoning is correct, but in any case, it
should be reasonable to not need to prepare edge extend surface for the
surface which could never been drawn.

Bug: 244260394
Test: atest ActivityLifecycleTests --rerun-until-failure 10
Change-Id: I31a36398bec86cb439396ad93108487cd6479139

Former-commit-id: e252ffde09a55e8fc0680ea8a9d2281f781b8b1d