Correctly handle null ProxyInfo in Ikev2VpnProfile
This change ensures that null ProxyInfo instances are honored
and reflected in the actual VpnProfile.

Ikev2VpnProfiles incorrectly set ProxyInfo as an empty instance
when provided with a null ProxyInfo. Other components in the
system sometimes have different behavior if presented with an
empty ProxyInfo instead of a null one.

Bug: 244281603
Test: atest FrameworksNetTests
Test: Manual testing with VPN validation enabled
Change-Id: I67ee2aa5c96dc22ba8fcd2be3fd4c090dec20697

Former-commit-id: 29348dbfca787235ae1eabbe1a8b7b262187f197