Fix crash where trim memory resulted in HWUI deleting the surface.
When HWUI is given the TRIM_MEMORY_COMPLETE signal it deletes all
of the associated GPU resources. In doing so it also deleted the
GL/Vk handle to the the ANativeWindow. So if the app attempted to
recover and draw new content it would crash because we no longer
had a surface to reinitialize and render into.

Test: atest hwui_unit_tests
Bug: 233388563
Change-Id: I54f3f0dadc3dfc89d2b1f567d36611f6a966c4f0
(cherry picked from commit 10568cc9611acb6c55c0694e4e8cfa739fcf0a48 [formerly b830772f7861207924b2a255a6ab9178befa664f])

Former-commit-id: 289380ff2b807e079724021294fcc783ed8ed46a