Introduces requestBaseStateOverride test API
Adds new requestBaseStateOverride on DeviceStateManager
to be used for testing, to provide a similar behavior
to changing the physical configuration of the device.

This is necessary now that the requestState API is being
used for features now, we need a test API that will
be a closer simulation to physical device changes.

Bug: 234336979
Test: DeviceStateManagerGlobalTest
Change-Id: I6ad9e799329f16521aa3c13b248eed762fe0b121

Former-commit-id: 7f8d5fd4514f812f0d6113fce203dc720a28c1eb