Fix a small bug where desktop mode changes might not be applied
Currently the window container transaction to apply desktop mode
would only be performed when the setting changes. If desktop mode
is already enabled, there wouldn't be a setting change event for it
being "on" after a reboot.

This CL updates the code to apply the wct in DesktopModeController
onInit, if the setting is enabled.

Test: manual - enable desktop mode via the tile
             - reboot the device
             - open an app
             => app should be in desktop mode
Bug: 241944030
Change-Id: I15df10dd275ef62df2cdcf7aaf22db526a842e97

Former-commit-id: c063e246ce294b05e5795dfa250eaf3b42f718cf