Fix cancelSynced cleanup: stop the vibrator.
Currently, the VibratorController is not updated when a synced vibration
is cancelled. Consequently, it stays "isVibrating" when the vibration is
finished. This is fixed by doing the same steps cancellation of individual
steps failure for the triggerSynced failure too.

Add new tests to ensure that vibrators have stopped vibrating, and to
stop the looper thread (this is how the issue was found).

Bug: 244428021
Test: atest
Change-Id: I67ffeb170ce28e1f65227079185f1cda75636a14

Former-commit-id: fb2fc1f0c733eefbe59024207493eca1d70beb3c