Reorder media player cards
Checking that the updated media player has data that isPlaying control
set to true and that of the current visible media player is set to false
before reordering. Also, checking added media player is active. So we
make sure we scroll to the current playing media player card. Media
players will need to be reordered later, if these conditions don't meet.

Bug: 234639220
Test: atest MediaCarouselControllerTest
Change-Id: I242c7d800a11d0525f9b4c525fe4885cbffd60d1

Former-commit-id: 460561d92e02f90110f15a8c23eebf211ad8e0cf