Handle embedding startActivity Intent when calling from application
Before, we only resolve the Intent if it is called with Activity
context, which skip handling cross-process embedding if the
startActivity is called with Application context.

Now, when startActivity is called with Application context, we will
check if the Intent should be embedded in the known Task in the
organizer process.

Bug: 244411454
Test: atest WMJetpackUnitTests:SplitControllerTest
Change-Id: Ic1ee4f23330b4ce35888c0eabd1ee8eb13d13140

Former-commit-id: 08966df3e830280086f43eee618932cc1d4478ba