Temporary fix on isUserVisibleOnDisplay(userId, displayId).
WM will use this API, but the current implementation would break
the behavior on non-passenger displays (like cluster, display, or
virtual displays).

So, this CL is (temporarily) changing the API to always return true
for the current user (and its profiles), regardless of the display;
the long-term solution would require integrating UserManagerService
with DisplayManagerService for displays management.

Also added a UserManagerInternal.isUserVisible(userId) method (which
will be used by WM during Activity transition).

Test: adb shell am start-user --display 42 13 && \
      adb shell cmd user is-user-visible --display 42 0
Bug: 244644281

Change-Id: I5330016e0123e85c0f2ab3fc24a4f683994fdfe8

Former-commit-id: 4ab0f4fe43a31487dcc412764baf8d7226334c6a