Fix ActivityEmbedding Shell transition animation with reparent
With legacy transition, we capture a screenshot of the reparent child,
and attach it to the parent surface (in WM Core), and then animate the
parent in handler.

With Shell transition, we also need to do the same (in WM Shell) to
screenshot the reparent child, and animate the parent using the child's
change info. Otherwise the child animation will be cropped.

Bug: 207070762
Test: manual
Change-Id: I52b5c839343864b4179e5c35c33343bc282bc189

Former-commit-id: c72e0d110f02367cf37b399a007f9be61aefbfc2