Improve AmbientState test coverage
AmbientState is a critical state holder for system holder but does not contains unit tests. The tests here are extracted from the attempt to convert AmbientState to Kotlin, and introduce a basic unit test coverage that we can build upon to increase changes confidence. Because AmbientState is a simple state holder, the tests are mostly checking property modifications.

We did not include test for any property that does not contain a custom getter or setter.

Test: atest ~/tm-qpr-dev/frameworks/base/packages/SystemUI/tests/src/com/android/systemui/statusbar/notification/stack/AmbientStateTest.kt

Fixes: 242754590

Change-Id: Ieb95b9845ee044ae293a3934b9a688c879c64249

Former-commit-id: faeaa3104957c5fa1f60746caf13516df4d22c61