Remove unnecessary calculation of display configuration
Since hierarchy unifying, ideally the configuration is already
synced if there is no orientation change. These code still remain
because the display config change from decor window relies on it:

Since [1] and [2], the height of navigation bar is no longer read
from resources. The height can be changed according to the layout
parameter. That may cause the display not to compute the latest
configuration when display density or navigation mode is changed.
And the display config change will happen at random timing when
calling updateOrientation (usually from resuming activity).

But after [3] and [4], the new display config can be sent if the
decor insets provider windows are changed. So now these code
are redundant, especially the computeScreenConfiguration is not a
cheap method, and it will be called twice from resumeTopActivity,
which adds extra unnecessary latency (20%).

[1]: If8ec79b3cb3b989eed578c44e8749b13fd2e8592
[2]: I04af809b5ff1ea43772a41d36ed6d97fb80ee02e
[3]: If966bcc8125300d47d5cd631f7db17ff027e5261
[4]: If6c28a89939372e6362b92f469ef7d793e8ec215

Bug: 234585256
Bug: 159103089
Test: CtsWindowManagerDeviceTestCases
Change-Id: Id7b7f4f7a647edd51b4d3255131eb2052f403784

Former-commit-id: d66df0deb4462b1345291b8513383d0d4b638034