Fix alpha fading of NotificationShelf  in LS
While we are on the LS and we close the Shade, the NotificationShelf
alpha was not animating. Now we share the same logic for calculating the
alpha values for the ExpandableNotificationRows and the
NotificationShelf.

Added new testCases to StackScrollAlgorithmTest to check an intermediate
state of the animated alpha values. Previously it was only checking the
end state, which is 0.0f.

Test: atest NotificationShelfTest
Test: atest StackScrollAlgorithmTest

Fixes: 237339523
Change-Id: I97b734d296397b6e2587512ff96ab62d976b6192

Former-commit-id: 3e531f56bf0f1a87e3d509f9c1643f847afa8600