Remove parameter from stopVpnRunnerAndNotifyAppLocked()
The current design makes the caller can pass any package name,
but in fact, stopVpnRunnerAndNotifyAppLocked() only cares about
the current prepared package. So if the caller pass a package
name which is not the current prepared, the method will not
work. Remove the parameter and use mPackage directly to prevent
the caller misuses this method.

Bug: 236409954
Test: atest FrameworksNetTests:VpnTest
Change-Id: I9297bf97fdcd7cef8ab0dadff3e6c05bf77fc667

Former-commit-id: 737d5033cc0db91c3e5359792f2f86e7d0810cfc