Disable intent filter enforcement in startActivityFromRecents
Due to the fact that the resolved type of an intent is not available
during startActivityFromRecents, the intent might not match the intent
filter, making the activity un-launchable if the enforcement is enabled.

The solution in this commit is a nasty hack for T; a proper fix should
be implemented in master.

Bug: 240373119
Test: atest CtsContentTestCases:PackageManagerTest
Change-Id: Ic88c728373d86dfddd6198600a05c696112048da

Former-commit-id: 5e6409d64fd0af1c8921b52a480d4b2c2b270b76