Make ATMService#mUiContext mockable
Some vendor-specific (lightweight) unit tests use the real class for
ActivityRecord but mock ActivityTaskManagerService (e.g.
ArcLaunchParamsModifierTest).

The problem is that ActivityRecord directly accesses
ATMService#mUiContext to get the DevicePolicyManager service and the
property itself is marked as final, so this cannot be replaced with a
mock and instantiating ActivityRecord fails.

Bug: 244515960
Test: atest ArcLaunchParamsModifierTest
Change-Id: I2fc015d9f3dc4ce88607355b8e2a2849c2ac7fb4

Former-commit-id: 4121d6bd7d7d23fa6dccaa73f53ff20884737f4a