Log global JobScheduler state changes.
Log state changes of constraints that apply to all jobs. These
constraints only need to be logged once (when they change) instead of
for each individual job and allow us to analyze behavior when
JobScheduler believes the device is in various states.

Keep the individual job constraint change logging off.

Bug: 129954980
Bug: 242611030
Test: `statsd_testdrive 514`
Change-Id: Idf331749601db6ed6f767d5649667519046208ae

Former-commit-id: 367e535ed110224f9bb16067515237c472e82d08