Check if caller is suspending package before unsuspended broadcast
The code falsely assumed the caller unsuspending a package was the one
previously suspending it. This checks that the caller actually is the
only suspending package in the map.

Also adjusts some test callback infrastructure to make it easier to
use and read.

Bug: 232319688

Merged-In: Ie06df12ff3da6cf209fb3149e6aa80b0e63ce56e
Change-Id: Ie06df12ff3da6cf209fb3149e6aa80b0e63ce56e

Former-commit-id: d161411c4450334b6af1475448798bcaf1ecaa36