Enable a shortcut selection test.
Enable ChooserActivity test that involves clicking on a shortcut.
A test can specify a new test override,
ChooserActivityOverrideData#directShareTargets lambda, that would return
a pair of a shortcut type and a collection of shortcuts, that would be
treated by the test activity as an output of the shotcut loading logic.

Bug: 215699869
Test: atest FrameworksCoreTests:ChooserActivityTest
Change-Id: I55da648585bd26374ede30dd5b45ed4af81d688e

Former-commit-id: 80f88eb6f01f2411da6a7bdfaeb9a561c1761ea1