Making the config_screenThresholdLevels a float array to fix the issue where it was always greater than all possible brightness values
config_screenThresholdLevels is an integer array, which is internally compared with brightness in [0.0f,1.0f] to get the right index on which the brightening and darkening thresholds are to be applied. This means no matter what we have in the array, it will always be the first index that will win, and we will always apply the first brightening and darkening thresholds. To keep the settings backward compatible, we still accept the integer array for this threshold, and divide those thresholds with the scale range(255).

Bug: 245746705
Test: Manual
Test: adb shell dumpsys display | grep HysteresisLevels -A 5
Change-Id: I177d7a1aa48bdfb3385780f0eee7af689755efb4

Former-commit-id: 8d81aca23a54414d0729b216b3e75b213371f340