Part 2 of removing shade blinking while moving heads-up notification
While touching HUN we're immediately setting some height of shade and in split shade - because QS are expanded by default - that causes QS height update as well which makes shade expanded for fraction of a second. Solution here is to not update QS height at this stage while HUN is visible but the rest of the shade should not be.

Test: Show HUN, try dragging it up or down and see no blinking of shade
Fixes: 234940017
Change-Id: I6dbaf211d15779573d89bd1bd1c87201207a5e71

Former-commit-id: baeac78271a14079d62cf6c132fba233942af58d