Add large screen clock motion - method 2.
When the clock moves on a large screen (e.g. when locking/unlocking the
device), this adds a subtle offset for each digit of the clock.

This hooks a new transition in to the existing clock move transition,
which just forwards progress through to the actual clock. This also adds
a feature flag (210) to enable/disable this, since it needs a little
more polish yet.

As above, this needs more polish (tweaking the time and/or easing), but
this at least gets something moving to start with :).

Bug: 199051139
Test: Manually verified.
Change-Id: Icec0bcfdb431d24b9405c5dfe8993016f687cac5

Former-commit-id: 82cec88397fb885a3d52f24162c9ca7725c91a06