Animate all targets to accomodate swiping during split screen.
These now scale out from the middle of the screen together,
instead of only one being visible and the rest jump-cutting in.

This should also obviate the need for the root task fix, since
we now animate all tasks WM gives us anyway. I updated b/232852578
to reference this CL in case that inadverently regresses something,
but it should not!

Fixes: 206741900
Test: added new keyguard test (atest SystemUITests)
Test: manual on tablet + regular device w/ no split screen
Change-Id: I7db8cda87b056970d9c3a6a383fb996489fd793e

Former-commit-id: 8bbe39d6854267c433f63f84b4354545d55fd63b