Add a way to disable non-wakeup alarm deferral.
AlarmManager normally delays non-wakeup alarms when the screen is off.
However, tests that depend on components that want to use non-wakeup
alarms while the screen is off will be flaky with this deferral. Make it
possible to disable the deferral for tests. We may also decide to
disable this behavior altogether, so use DeviceConfig to allow
experimentation.

Also fix long comparison overflow issue in AlarmQueue.

Bug: 245987711
Test: atest CtsJobSchedulerTestCases:FlexibilityConstraintTest
Test: atest FrameworksMockingServicesTests:AlarmQueueTest
Change-Id: I35f1299d551eb7099adfecd13449f53fef88de02

Former-commit-id: ae9ea5f3d9f9a2193bba679dfd347191591e79f4