Move PiP swiftly in response to IME with keep clear areas flag enabled.
If the keep clear algorithm is enabled, PiP window will avoid the
keep clear areas and move away from occluding keyboard.

When IME is shown, we trigger the movement right away, for a smoother
experience.
If there is another animation ongoing, we delay moving PiP again, so
that it's more smooth overall.
Finally, we always apply gravity bottom with keep clear areas flag on,
so that it goes back down after IME dismiss.

Test: manually, http://recall/-/blXo3gy66k4fDYh7gFkrpe
Bug: 183746978

Change-Id: I32c52e4b6dd3c083b2168fda6a8b3b13facdac69

Former-commit-id: 62f6fe2294dfd866beb62fdf401f27d0c7966baa