Extracts interface from UserSwitcherController.
Our goal with this refactor is to provide a differnt implementation of
UserSwitcherController that uses modern architecture inside itself. To
achieve that, we need to be able to provide a new implementation
alongside the existing implementation and switch between them based on
our feature flags.

This CL extracts an interface our of the UserSwitcherController class,
also named UserSwitcherController (to minimize changes to downstream
customers of this class) and moves the current implementation to
UserSwitcherControllerOldImpl.

As a side-effect, we also had to move the old
UserSwitcherController.BaseUserAdapter out into its own class,
BaseUserSwitcherAdapter.

This CL was done almost entirely automatically using tools provided by
the Android Studio IDE. There are no logical changes in this CL at all.

Bug: 246631653
Test: Manually verified that user switcher works properly in
full-screen, from quick settings, and from the bouncer.

Change-Id: I383f1bb9147c7afc3047bcf2ad7422b7c6894821

Former-commit-id: e112db8ac61fca2a5a1d1c73b56b2223d93f816a