Make transit to home activity always be close transit
Previously, we used the wallpaper transition when going to the home
activity. The idea was that on devices that have no wallpaper, this
is safe. However, the check for the wallpaper was incorrect.

A better approach is to detect when a task open transition is
happening and check if it's the home task that is being brought to
front. If yes, then we use a close transition, closing the
previous activity to show the home activity

Bug: 237371677
Bug: 220024691
Test: m && manually check that keyevent HOME fades out the top app
Change-Id: I46ba3d363b581f07e153e032689261821cbbe9e4

Former-commit-id: b576faec5c70f661c4d392c4a73baa7974341378