Track order of visible media players
Create a new map that track the order of visible media players before
they are reordered as the media carousel is visible to user. Media
player map that was being used before is not consistent with UI during
the time the carousel is visible to user. So during this time we need to
keep track of the order.
This also makes sure the media carousel scroll to the right app media player when user selects one of the recommendations.

Bug: 248340474
Test: atest MediaCarouselControllerTest
Test: Checked order of UMO cards in carousel.
Change-Id: I549535125b251b03740a387cc4a1cfae91f4f4f5

Former-commit-id: 49f68af1be448ef36d434dabb9e217e6a6643bb5