Introduce system clock time confidence
Add metadata for system clock time that indicates confidence in the
time. Any explicit action to set the time will set the confidence high,
but this information is (deliberately) lost during a power cycle as the
clock is (sometimes) volatile.

This confidence will be used initially to support setup wizard and other
initialization flows that benefit from knowing whether the time needs to
be checked by the user. This lays the groundwork for future
improvements.

As part of this change, bug report logs can now contain information
about how the device's time was set by things other than the time
detector. This will help with debugging.

Bug: 236612872
Test: atest services/tests/mockingservicestests/src/com/android/server/alarm/AlarmManagerServiceTest.java
Test: atest services/tests/servicestests/src/com/android/server/timedetector/TimeDetectorStrategyImplTest.java
Change-Id: I1104e64778f57ef3d1716e9aba1736c4fd001ef7

Former-commit-id: c4ba2d9ce2d5cb4bf98e652ec3803068dcdd31d6