Add id to InputDeviceDesc of VirtualInputDevice
Virtual input devices did not store the deviceId as this was not needed.
This change stores the deviceId during the device initialization (deviceId
is given to InputDeviceListener callback, which is used to wait for a
virtual device to be initialized and configured)

The deviceId is needed to restrict an input device to a display group.

Bug: 244516415
Test: atest android.hardware.input.cts.tests.VirtualKeyboardTest
Change-Id: I982191d543556e2728e41d6da01ab1a631e50e1a

Former-commit-id: b03a7b7d075d10ffd21eab2960e03237cabbda32