Introduce ImeVisibilityApplier
With go/new-ime-visibility-control-u, this CL introduced
ImeVisibilityApplier interface to abstract the implementation of
applying IME visibility with adjusting IME z-ordering for aiming to
stablize IME z-ordering control.

Note that this is the first CL with
- Refactoring part of IMMS#{show, hide}CurrentInputLocked logic to a
  default implementation class of ImeVisibilityAppler
- Clean-up IMMS#{mShowRequestWindowMap, mHideRequestWindowMap} with
  replaced by ImeVisibilityComputer.WindowState#setRequestImeToken

Will keep update follow-up CLs for clean-up applying IME visiblity and
adjusting IME z-ordering stuffs.

Bug: 246309664
Test: atest CtsInputMethodTestCases
Change-Id: I410a29ce4a4e27b2ffa66c9d090d969eb43d0e36

Change-Id: I12d99c15ae0d8965a21406d2495ce5cb18afaea0

Former-commit-id: baebf0fecf959b93981fbf43e6c7fa156cdbdbea