Moves enforcedAdmin into UserRecord.
In an effort to move logic out of the old UserSwitcherController
implementation, we are moving the concept of an enforced admin for user
actions that are disabled by an administrator into the actual
UserRecord itself.

This CL has no logical changes beyond moving where the state is being
held to a more common one.

Bug: 246631653
Test: Manually verified the old implementation of the full-screen user
switcher, the quick settings button, and the lock-screen user dropdown
all still work properly.

Change-Id: I8b6bd8f3424fe531aa4889f530968f6f48fec536

Former-commit-id: 963f710d4c04febb860ae9daf06a4cea44b4720f