BroadcastQueue: skip disabled apps/components.
The OS can request that a package, sub-components, or an entire user
be "cleaned up", typically when something is disabled.  This change
implements this in the modern broadcast stack, along with tests.

We're treating "cleaned up" recievers as being "skipped", since we
can't change the indexes of BroadcastRecord.receivers after they've
been enqueued.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: Ic4061288c72ac0528e9662cb973a189e4572163d

Former-commit-id: 05ec471715f32cccdc8de54eb5efac76dcd1b7c5