Add duplicated DisplayPowerController
We are creating a copy of the existing DisplayPowerController so that it
can be used as a base to apply the refactoring ideas. This is crucial as
the old DisplayPowerController will continue to behave as it does today,
minimizing the risk of any bugs that might get introduced.
Reviewers of DisplayPowerController class are expected to ensure any changes happening
in DisplayPowerController are also accommodated in
DisplayPowerController2.

Bug: 241307688
Test: Manual
Change-Id: I7682996d4e47cece6cdfc3e4e6e539677e85496b

Former-commit-id: 433cfd2d4cf77ad748f40284121e52b4f8b39437