BroadcastQueue: active counters, usage reporting.
Other parts of the OS (such as OOM adjust calculations) rely on the
number of active broadcasts maintained in ProcessReceiverRecord, so
this change adds them to the new stack.  The new stack doesn't need
to do membership tests, so it migrates towards a faster reference
count approach.  Tests to confirm behavior of both stacks.

Dispatch usage events expected by PackageManager and UsageStats and
skip broadcasts to apps in full backup to match previous behavior.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: Ia3faf26d01ec9d662d2377a0b47b28914cfe8e03

Former-commit-id: 6ba529798182ca9e5a979db01b5b35f244531281