Extract appop restrictions
Extract all (or most) app-op restrictions behavior out of AppOpsService
and into LegacyAppOpsServiceInterfaceImpl.

There are two types of app-op restrictions: global restrictions and user
restrictions. Currently, the code/variables for both of these features
resides in AppOpsService. As part of ongoing refactoring work, we want
to move this functionality into another class and provide an interface
which AppOpsService can use to access it.

Test: atest AppOpsLegacyRestrictionsTest
Bug: 240200048
Change-Id: I18121136c0855d72b3b530b68dce2e2de4d40d16

Former-commit-id: 7699ecdf364d67c9f64a8d31a3a9a76c92484882