BroadcastQueue: tests for ordered broadcasts.
Verify that a normal ordered broadcast collects results from each
receiver in the expected order along the chain before being delivered
to the final result callback.

Also verify that an aborted broadcast stops dispatch to any further
receivers in the list, and that it's still delivered to the final
result callback.

Since the modern implementation isn't finished yet, use assumptions
to skip testing it; we'll enable them once implemented.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: I651e6c9f4b08822f0f222ea02b7ea9a89eb81da9

Former-commit-id: 6e559903d5d8a73268e7539cbf9d399c3bf3371e