Updating the cached resource configuration if ViewRoot is perserved
When configuration changed, an Activity was relaunched because the
new layout size is smaller than old layout size. However, the
mLastConfigurationFromResources was not updated when relaunching
Activity while the window is preserved.

If the same Activity was resized to the previous bounds and the
configuration changes are not cross the size buckets, the Activity
won't be relaunched. Meanwhile, the View#onConfigurationChanged
was not reported because `mLastConfigurationFromResources` was not
update-to-date and was the same with the new configuration.

Bug: 247013643
Test: resizing the app in split-screen
Change-Id: I392417e1e6cc8bb5e8be0006b3b81385c3ef41aa

Former-commit-id: 615928575a45e15d0583329bb99b39dc915b5e88