Cancel AodTimeoutAction when finger lift.
If AodTimeoutAction is not cancelled when finger lift and user try to
unlock the screen again, AodTimeoutAction will still send finger up in 1
second. So the second unlocking will be terminated by fake finger up
event. But the user's finger haven't leave the screen, the UI will show
third unlocking immediately. So AodTimeoutAction needs to be cancelled
when finger lift.

Bug: 247048663
Test: Check UI works properly
Test: Run atest UdfpsControllerTest
Change-Id: I1a4a0038e514ca1045db6894461e790c67cb0084
Signed-off-by: Wendly Li <wendlyli@google.com>

Former-commit-id: 6b1747884759016c2903bc276d32c0ec391d6c81