Eval KG showing, occluded and visibility together
Keyguard visibility is based on the keyguard showing
and occluded states, but the KUM state changes were previously
separated so that KG would respond to a occlusion change
but not have the newest visibility, so it would
inaccurately update its fingerprint listening state. To avoid this
race condition, this CL consolidates KG showing, occlusion and
visibility changes to a single setKeyguardShowing(showing, occluded)
method. The method updates the KUM state and then sends the change
out to listeners.

This CL also updates some of the vocab used. In the future,
keyguardVisibility from a KUM perspective means that the keyguard
is showing and NOT occluded.

Test: atest SystemUITests
Test: enable a screensaver and successfully authenticate from the
screensaver with a non UDFPS fp sensor
Test: double press power button to the camera (occluded state), and
see that rear FP will stop running. Can still auth when accessing
the photo gallery.
Test: double press power button to the camera (occluded state), and
see that UDFPS affordance isn't available. Can still auth when
accessing the photo gallery.
Test: Setup a timer to go off, see that going to/from the occluded
activity updates the occluded & keyguard visibility state correctly
Bug: 248089638
Fixes: 247580913

Change-Id: I3478be2d304a42beebdb6a4baaf2e571d74e61b1

Former-commit-id: d2dfe82bd8f4953184938e0e73a263e6535b6af7