Fix screenshot swipe dismiss logs
Currently we log a "swipe dismiss" event whenever we animate to the
side, even if it wasn't caused by a swipe -- e.g. if the UI times
out, we log both a "timeout" and a "swipe dismiss" event. This
change fixes the logging logic so that we only log swipe dismiss
if it was caused by a swipe/fling, as well as adding a couple
of missing dismissal logs in edge cases.

Bug: 252823661
Test: atest DraggableConstraintLayoutTest

Change-Id: I437de637e606f4d2fe806ca6dd6a397b4ca372c6

Former-commit-id: 3c7ecee13c146cd4423cc648aa0ae1fa006b7d83