Removes isAccessibilityTool from dynamically-configurable properties.
AccessibilityServiceInfo#setAccessibilityTool() is already a hidden
method but reflection techniques may be able to bypass this. So, we
shouldn't allow this property to be modified at runtime by an
a11yservice after the service info is parsed from the XML.

This field was originally added to configurable properties to allow
modification during CTS tests.

This change replaces dynamic configuration with a flag in UiAutomation
that is used when connecting a new UiAutomation instance. CTS tests
which wish to change isAccessibilityTool mid-test can reconnect their
UiAutomation with a new flag value. The flag is @TestApi since it is
unlikely non-CTS developers will want it for their own tests.

Fix: 248335757
Test: atest AccessibilityEndToEndTest
Change-Id: I6ca3d9e6fb94207897cdfc94925bbd4d54348a29

Former-commit-id: ae6b03f8e1400f0d84aae8572e6556ba97b5fea1