Dismiss action menu when item is clicked.
There is a bug in the full-screen user switcher UI. The action menu
(accessible from the "Add" button) does not dismiss itself when one of
its items is clicked. This issue is only present in the new
implementation and no in the old implementation. This CL fixes that for
the new implementation.

Bug: 243844359
Test: Manually verified that the dialog is now dismissed when one of its
items is clicked.

Change-Id: I051dad133d94d454836f7c6868d57b7085ee1412

Former-commit-id: 1cb81f64393c494af7923bfea8f6df2fa6aa112c