Update ASM activity start rules based on new metrics.
This change updates the rules introduced in ag/19532978 based on the
findings in go/asm-buckets to reduce impact on apps. Namely, the rules
have been updated to:

1. Within a task, only an activity matching a top UID of the task can
   start activities
2. Only activities within a foreground task, which match a top UID of
   the task, can create a new task or bring an existing one into the
   foreground

These rules are not yet final, and may be updated again once we have
more data from droidfood.

This commit also starts to allow launches for a subset of the BAL
exceptions, even if they violate the above two rules. See go/asm-buckets
for the full list.

BYPASS_LONG_LINES_REASON: Auto generated enum names
Bug: 246738504
Test: atest ActivityStarterTests BackgroundActivityLaunchTest
Test: Manually tested scenarios from go/asm-buckets.

Change-Id: Ia4bc2eeb0006c94c006072a841ebba7d3bd7b99b

Former-commit-id: 464d37173f665ae0a7fa8ef0523a0f2cef5e68c6