Move MediaProjectionGlobal API to DisplayManager
Add a static method in DisplayManager that allows the caller to create a
VirtualDisplay without having an instance of DisplayManager (which
requires a context). This can only be used from a process that has the
CAPTURE_VIDEO_OUTPUT permission.

Also added RequiresPermission(CAPTURE_VIDEO_OUTPUT)
annotation for the new API.

Test: DisplayManagerTest
Bug: 247776252
Change-Id: I1eb94ab1063d89860cdfb2a58d4653d9a77eec98

Former-commit-id: c6f10b42d9d905da3a8e41033b1233b9212274e8