Making sure QS panel is always visible when expanding split shade
When we swipe quickly on launcher to expand shade sometimes we go through path where notifyExpandingFinished() is called at the beginning of the movement - when focus transfer happens and expand() is called. This sets `mQsExpandImmediate` to false but it should be always true when expanding split shade.
Fix is to move resetting `mQsExpandImmediate` to onPanelStateChanged reacting to OPEN/CLOSE which should be called only once and only at the end of motion.

This fix also coincidentally fixes b/237909902 as it was caused by not resetting `mQsExpandImmediate` after showing heads up notification, now it's done when we transition to CLOSE state.

Test: NotificationPanelViewControllerTest
Test: which quick flings on launcher try to expand and close shade many times -> notice QS always visible
Fixes: 248555255
Fixes: 237909902
Change-Id: I64e54765fe39182ff7f71d1d6407d3e38df051b1

Former-commit-id: 3e9e2d401c7c91ba4e98a235564d9117f876e018