Adds back using WindowState for magnification letterbox bounds.
This is a partial revert of
commit 5bfc2523f57c11b947992e16a84115961c776edb [formerly 35280c1dc4e9777d41d56f8a2cd991a577d3b7c8].
Magnification bounds still use WindowStates for its window iteration
instead of InputWindowHandles, so that change shouldn't have fully
removed WindowState letterbox bounds handling yet.

This change also includes a fix for the bounds calculation: the previous
logic assumed that the letterbox insets were negative instead of
positive. This new version reverses that assumption.

Example bounds for an app in landscape mode, letterboxed to the center
of the screen:
      app bounds       = Rect(663, 0 - 1677, 1080)
      letterbox bounds = SkRegion((0,0,663,1080)(1677,0,2340,1080))

Bug: 248059823
Test: use magnification on a letterboxed app
Test: log app & letterbox bounds, ensure they match expected values
Change-Id: Ic77474c416d0e7097428ea0303302958701e8387

Former-commit-id: 3c44c6926b8f81b075a15b887110b94694b3164d