BroadcastQueue: tests for failing starts, dead.
Add tests that verify the behavior of failing cold process starts
and handling of DeadObjectException.  This generally means we
skip the active broadcast and keep making progress on our queue,
instead of risking stalling out.

This testing uncovered a subtle bug in the "default" implementation,
where we'd clean up a dead registered receiver, but leave a dead
manifest receiver floating.  We fix this by giving both cases the
same treatment.

Add string caching to improve performance.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: Ic7ece84d6b0404d49b2fdf586f4574efdd4175fa

Former-commit-id: 5f31d431720afee05d10c77c618dd52e2cb8243b