Always report Activity configuration changes unless relaunch
Otherwise, neither the Activity#onConfigurationChanged is called,
nor the Activity is relaunched.

This CL is restoring the behavior to before commit 42b7273 landed.
The commit was meant to skip reporting the config changes to the
app because the app is not interested in handling the changes.
However, there are libraries like Compose that are still relying
on the callback.

Bug: 247013643
Test: AppConfigurationTests ActivityThreadClientTest
Change-Id: I552a5bccdb7862b0e903548e057d2cf94ac28c83

Former-commit-id: 44df9d80a68a8955729da1ae84b4ef19e729d0ed