DisplayPolicy: Check that window provided insets
This is the fix to a problem in WindowState.removeIfPossible(boolean)

The core of the problem is that WindowState.removeImmediately call removes
providedInsetsSources, but these are important for
WindowState.providesNonDecorInsets to check if the window provided
insets.

This solution makes the call to WindowState.providesNonDecorInsets
before calling WindowState.removeImmediately

The result is that sendNewConfiguration consistently triggered during
removal of navbar.

Bug: 211763738
Test: atest WindowStateTests
Change-Id: I2b92688437ffab02f4aee8519eb2a8e3deeb6484

Former-commit-id: 097829d3492a90e2613be00ae0c3d77d20875613