Platform Telecom Transactional APIs
Telecom Transactional APIs are defined as APIs that leverage
android.os.OutcomeReceivers.  The receivers are to be completed by
Telecom for CallControl opertaions and the Client for CallEventCallback
operations.  Doing so ensures the client and telecom are in sync with
call states.  Also, these APIs are more lightweight than the
ConnectionService way of starting a self-managed call.

bug: 249779561
Test: atest android.telecom.cts.TransactionalApisTest

Change-Id: Icb09b2874d599a40afca8b7e960b14ca1bca606d

Former-commit-id: 9bfb2434c5b04f1ce3acd7805a0dc571768579b2