Add CPU ABI fallback directly into PackageState
To simplify the external API, the raw package ABI won't be exposed,
only the overridden/device configured value available through the
PackageState object.

However, changing the default fallback behavior would also change
the value that existing callers get within system server.

It's difficult to audit whether or not each caller needs the fallback
or not, so the behavior was maintained with separate
getPrimaryCpuAbiLegacy and getSecondaryCpuAbiLegacy methods. These
should gradually be removed as the callers' needs can be identified.

Bug: 225395398

Test: atest PackageManagerSettingsTests
Test: atest ScanTests

Change-Id: Ibdf49d700ab6f21afa5a61c66f47dc2844b6cd45

Former-commit-id: c87545bf7d2ebb22cdb689b4a0540236b085d98b