Add `compatibility-tradefed` for `libincident_test`
All of a test's dependencies should be explicitly declared in the Soong
build module to enable infrastructure features that ensure correctness
and optimize runtime.

The `FilePusher` class which is defined in `compatibility-tradefed` is
required because the corresponding TF test config has a target preparer
that uses the class and its specific option, `append-bitness`.

Note that we would not normally need to add `compatibility-tradefed` to
a test that's part of a test suite (e.g. CTS, etc) since the dependency
is implicit. It is required in this case, however, because the test is
not part of any suite.

Bug: 220991354
Test: atest libincident_test
Change-Id: I41feeab36744e25c740f0c0cbf74973e6ff97607

Former-commit-id: ba21089dc9ebda077500d07f5481b0e1c991e7d0