Fix security dialog bug
This CL fixes a bug introduced by ag/19712818: when inflating the
security dialog content, we should use the Quick Settings context (which
is always using dark mode) and not the application context. Otherwise,
if we are in light mode, then the color of the text is the same as the
background and it looks like there is not content in the dialog. In dark
mode, everything was looking fine because the text color from the
application context would be light.

Bug: 247961387
Test: Manual
Change-Id: I0088206b8a3f5473f22d6f80e89a5dd07875c167

Former-commit-id: aeae9af5f5be0cd74e533401728e8baa3c6aa523