Introduce APIs to retrieve only carrier configs as needed
Current CarrierConfigManager APIs only allow to retrieve
ALL supported carrier configs at once. It may be as many
as hundreds of keys/values in one bundle which is not only
slow but also consumes unnecessarily more resources.

Clients study shows that the majority of the clients actually
only care about one or few of carrier configs.

The new APIs allow clients to specify a single or
multiple configs they care about. The performance profiling
shows that it is not only much faster but also consumes much
less resources.

Bug: 244087782
Test: atest CarrierConfigManagerTest
Change-Id: I436b9ad1d7e928d628c0103144cd345e8c4c4e3c

Former-commit-id: 365825e2c587d7c6a8e324a90a022f94b5ab2aeb