On notification update, take the newest value of shouldHeadsUp
When addressing only notification updates, this is a very small edge case that will only occur if multiple notification posts/updates occur in the same build cycle, *and* the value of shouldHeadsUp changes.

This change is also being made in preparation for updates to handle onRankingApplied, which may add additional cases where an update occurs to the value of shouldHeadsUp.

Bug: 248325248
Test: HeadsUpCoordinatorTest
Change-Id: I274c5f65d92fc151bacaacb21726445f1bf7fecb

Former-commit-id: 0d3b37cdce43ff87006d2ba533ff4242af39ee9b