Fixes bug in user switcher bouncer dropdown
The bug was caused by the blind readding of header and footer views to
the ListView that backs the dropdown popup menu. The bug pre-existed the
refactoring work but was hidden because there were fewer updates being
sent to the UI.

The bug means that, if the bouncer dropdown is opened quickly after a
user switch event, the dropdown will visually have larger gaps above and
below its items.

The fix is simply to check whether we've already added a header and
footer to the list view before adding them again.

Bug: 246631653
Test: the bug no longer occurs after the fix
Change-Id: I58d307d79a0c7804636f145df22fe3e8f8e134fe

Former-commit-id: d544909970e984e42547d67b405e4c19dd6d74e3