Use @EnforcePermission for IIMM#isInputMethodPickerShownForTest()
This is a follow up CL to my previous CL [1], which locked down

  IInputMethodManager#isInputMethodPickerShownForTest()

with

  android.Manifest.permission.TEST_INPUT_METHOD

permission.

After the original CL was committed the severity assessment was
performed again and in the updated assessment it was concluded that
denial logging was not necessary.  With that, this CL simplifies the
logic by using

  @EnforcePermission

annotation in the ADIL method definition.

Note that there must be no developer-observable behavior change in
this CL, and the security test that was added as part of the original
effort [2] still verifies that the method in question is indeed
guarded with "TEST_INPUT_METHOD" permission.

 [1]: Ie79a3e9d41ce22605ae083594d639c37d08b7def
      e67074ad51444e3a7df90523ecf01246bd8e8f35 [formerly b869c783808c4d4937727a1672b2fac81bace368]
 [2]: Idf907e3b762307696a3a7ca11470b0c44b9b7aa4
      3e1dd9d2797f818766105247e3634da525fec8e8

Bug: 237317525
Test: atest CtsInputMethodTestCases:InputMethodManagerTest#testIsInputMethodPickerShownProtection
Change-Id: Ib3f56b1ab1538742a2bf64fec435ced5b8f90bc2

Former-commit-id: aa29be8dda97b0159a7805363dd28a0467a2bd5b