Add AccessibilityManagerInternal#isTouchExplorationEnabled()
This is a follow up CL to our previous CL [1], which added a
method invocation of

  AccessibilityManager#isTouchExplorationEnabled()

from InputMethodManagerService.  The issue is that it is not clear
about which user we are querying.  In order to avoid future
confusions, it would make much more sense if we have a more dedicated
internal method in AccessibilityManagerInternal as we planned.

Bug: 207007377
Fix: 219056452
Test: presubmit
Change-Id: I99a1b6706c51709e82ca31449753b35919e7247c

Former-commit-id: 1dd7a063838d88092bb25915f795d353fe100cfe