Mark actions unimportant for accessibility when obscured by RemoteInputView.
To prevent animation jank, we don't actually hide or remove the action
buttons (like "reply") on a notification while they're obscured by the
remote input view (like while the user is typing their reply).

Visually, this is fine, but Voice Access still sees those buttons as
important. If the user speaks the action (like "reply"), Voice Access
will activate the buttons, but nothing visible will happen. If the user
has numbers shown, Voice Access will number the buttons, but the numbers
will appear to point at the remote input view instead.

To avoid this, we mark the NotificationActionListLayout as unimportant
for accessibility while the remote input view is visible, and return it
to "auto" important when the remote input view isn't visible.

Bug: 230297786
Test: atest NotificationContentViewTest
Change-Id: I50fdda3adabd0039ed84b3781c0d275fc176ef89

Former-commit-id: 152fc131c53fbf62f84612f5aafc6aace58f57ad