Optimize IMMS#setAdditionalInputMethodSubtypes() a bit for bg users
This is a follow up CL to my previous CL, which enabled background
users to call

  InputMethodManager#setAdditionalInputMethodSubtypes().

One this this CL tries to optimize is the fact that its implementation
is currently invoking

  AdditionalSubtypeUtils.load()

twice when called for background usres.  The double-call is happening
in

  InputMethodManagerService#setAdditionalInputMethodSubtypes()

itseif and inside

  InputMethodManagerService#queryMethodMapForUser() [2].

With this CL, the load operation should be happen only once.

There should be no developer-observable behavior change.

 [1]: Iac778d95846843504f6e00b3b939ef0794013b02
      965be9bb96e788b38ce62bf03c9d64d360cb4d64 [formerly 0b169d04c113684b0213a3a15bac9e6f0d77e515]
 [2]: I4187468076705ac597d680f2f5dc32d7b166da1f
      6d7e55ad2f56c2fa52c3839a46e7f29cfcb7c8ab [formerly b5f3d34def7b979e611ca4abc3f9aa4a071a0fdc]

Bug: 234882948
Test: presubmit
Change-Id: I6eb4201ebd69ca8821cbe74fc69356f68f9bcb5e

Former-commit-id: 9ddd4cd83eb91c65e3f2529b187cc66041f6efcf