Fixes bug where affordances are visible in AOD
The issue is that, when consumers are subscribed to the repository's
isDozing flow, they get an initial value of false and are not updated
until after the doze state is changed.

The reason this matters is because the subscribers in this case are the
quick affordance button views themsellves and they subscribe and
unsubscribe based on their visibility. Hence, they get resubscribed
often, seeing the false value.

The fix in this CL is to give the true initial value to the newly
subscribed instead of just false.

Fix: 249084394
Test: Manually verified the fix. The attached bug doesn't reproduce
after the fix. Also tested with "Control from Locked Device" being set
to true and verified no harm done. Unit test case added.

Change-Id: I93369cc559af4ce266bfaec9372c748c8a353466

Former-commit-id: 3b251021c6324881c1622ebc582638f337bf5362