Fix QSPanel horizontal layout issue after unfolding
QS (Quick Settings) layout could be horizontal  or not. When it is horizontal, the media player is displayed on the side of QS.

 This is based on orientation, split shade state, and media player visibility. However the configuration changed lister was triggering the update only after orientation changed, neglecting split shade state.

When a foldable device that is in the state "folded and landscape" is unfolded, the orientation stays the same, despite the split shade state changes.

QSPanelControllerBase was updating the QS layout only if the orientation changed, neglecting other condition changes.

Fixes: 216244185
Test: CorrectQSAfterUnfold QSPanelControllerBaseTest
Change-Id: I7f1444adf477f75c267a48bf8f420d3507f82709

Former-commit-id: 3e6a8bcf0d78f2ed6fc4c298c5d25ace21c11ac8