Fix OverrideTaskTransitionTest w/shell transitions
The custom task animation test previously asserted that the Launcher
must become invisible when the transition starts. This is either
irrelevant or incorrect depending on the splash screen behavior:
- If the splash is opaque: the splash is covering everything, so the
   Launcher visibility does not matter.
- If the splash is translucent: then the Launcher being invisible would
  mean there's a flicker.

This change removes that assertion and replaces it with assertions that
verify that the override custom launch animation is played (the
show->hide->show animation).

Bug: 249026022
Test: atest OverrideTaskTransitionTest
Change-Id: I107839e9558de61a546ea67c0ed85aab6879b53a

Former-commit-id: 35307f822b3e1b61fc5ffc35d5478209ed25035c