Allow Android Auto to move to trusted.
Seems like there is a bug with how TrustManagerService deals with
automotive now that there is the possibility for renewable trust,
preventing the automotive companion from being unlocked by the phone. To
fix this, we'll update canMoveToTrusted to also allow automotive
trustagents to automatically move to trusted (if the right other
conditions are met).

Test: Manually for automotive
Fixes: 250703401
Change-Id: I906be77a2f09582095d24797aebdb0eda420eee6

Former-commit-id: 2fa61e04624ec9ccbfb9945b7b911ca564b6b6fb