Create VelocityTracker#getAxisVelocity API
This API allows clients to query the velocity over a given axis. This
became necessary as VelocityTracker starts supporting axes beyond X and
Y axes. The immediate reason why we are adding this API is to support
AXIS_SCROLL velocity. When other axes are supported in the future, no
API surface change is needed, as this API takes a target axis. See the
design doc go/velocity-tracker-for-axis-scroll for more discussions for
the API signature choice.

Bug: 32830165
Test: CTS tests
Change-Id: Ib150ea8177c1f207c277d7889ed71515a29c64d3

Former-commit-id: 93906cabb22f55e36666889a58b0ab7aba8f4fe0