Remove InputMethodManagerDelegate#getLockObject()
This is a follow up CL to our previous CL [1], which introduced

  InputMethodManagerDelegate#getLockObject()

to clarify that "mServedView" and "mNextServedView" need to be guarded
with

  InputMethodManager#mH

as a short term workaround.

With this CL, above two fields are moved back to InputMethodManager
like they used to be before we introduce ImeFocusController [2].

This is a mostly mechanical refactoring.  Hopefully there should be no
observable behavior change.

 [1]: I9c072b829d1db1e68b65e766d764ee71cb16e6a2
      b059e2078bf7078b7298da1bd919db74572574aa [formerly 31e467846378c12ca51c92cc4cc35e59d495288d]
 [2]: Ib455704fe1e9d243f93190a84f230210dbceac2a
      6fa5c53fe7e8207efa2fd82566e0d1e9d640ce16 [formerly 970d9d2e0c979cf9a0ff0a79ef49044ed1363d4f]

Bug: 244504062
Test: presubmit
Change-Id: Ie3446e1d0f62c489195e31fc7f073020884d6cae

Former-commit-id: ff975479b4f60fe1c599826dd6f42f24a42883f9