Add BackupTransport#getBackupManagerMonitor
This allows the framework to ask for a monitor to use for backups that
are initiated by the framework (i.e. scheduled backups). Previously
the only way to pass a monitor in was through the
BackupManager#requestBackup API.

I also moved BackupManagerMonitorWrapper to a separate class now that
it's shared by three different classes.

Bug: 252763821
Test: atest CtsBackupHostTestCases
  Also manually by implementing a simple monitor inside LocalTransport
  and calling the new getBackupManagerMonitor.
Change-Id: I84d56a3e80d6c94f9a327f7d71876d1bc27fe5ac

Former-commit-id: f7fbe355edae57849cb8acdee67d30bf50340139