Fix rounded corners cropping for letterbox
When using two apps in horizontal split-screen, taskbar is shown, and
the bottom app is "portrait-only" and letterboxed, this bottom
application is partially covered by taskbar.

This is because, the top of taskbar is calculated relative to the whole
screen, while cropBounds.offsetTo(0, 0) makes cropBounds coordinates
relative to the lower part of the screen.

Reordering the statements of "adjustment of crop bounds" and "offsetting
coordinates" fixes this issue

Change-Id: Id1ac9252fbf102e321a89330aefda9ca2624fe66
Test: atest SizeCompatTests
Bug: 245521412

Former-commit-id: 7a7c6d387fdb04b564f48d014eac97392fc94e76