SurfaceControlViewHost: Expose transferTouchGesture API.
A key feature of SurfaceControlViewHost is that touches are given
directly to the embedded View hierarchy, however in some cases
this can interfere with the seamlessness of embedding. For example
if we are not clicking on an embedded element but in fact
intending to scroll its container.

Allowing transferring the current touch gesture to the host
provides a way to solve this problem. We expose this through
a SurfaceControlViewHost specific API in order to restrict scope
and avoid sharing of any new tokens.

Bug: 256993188
Change-Id: If94dbe4b862b6a159f2c0baee55dcb8c82c22e37

Former-commit-id: 716ac1c2a93970c96bbff7d067abf74c465b89ea