[Bouncer] Add manage users to bouncer.
Add a manage users option to the user switcher for the Bouncer.
This adds a isMangeUsers property to the UserRecord and makes some
changes around that.

If user is on bouncer while trying to go to manage users or add a child
user, bouncer will show a message to indicate that user needs to unlock
the device in order to proceed.

Bug: 241811308
Test: Added a unit test and manual test. Manual test consists of testing
with new userswitcher code flag on and off, testing non fullscreen user
switcher, user switcher activity, and user switcher in the bouncer. In
addition to testing the manage users. I've added the max number of users and tested add user and add child user. I also tested scenarios where we click on manage users and then exit bouncer.

Change-Id: Ib20c32914fd95eb891676028b5b588009b9e50a6

Former-commit-id: bb821889a84299c79878a54505283ef70bb8cf03