Introducing the notion of DisplayBrightnessController, and making it use DisplayBrightnessModeStrategies
DisplayBrightnessController is expected to handle all the changes that
happens to the brightness of the display. We have also introduced the
notion of DisplayBrightnessModeStrategies, and are making this
controller implement 2 different strategies - DozeBrightnessModeStrategy
and ScreenOffBrightnessModeStrategy. In the future, we will be
introducing more such strategies.

In future, DisplayBrightnessController will also be responsible for clamping the
brightness, and notifying the system about the changes

Test: atest ScreenOffBrightnessModeStrategyTest DisplayBrightnessControllerTest DozeBrightnessModeStrategyTest
Bug: 253226419
Change-Id: I6658ccf778274378194e158b4e9f19fb04dc1d3f

Former-commit-id: 5909714ea385afe509ecc138da88ccbefccd51a5