Allow seamless rotation on large screen with movable navbar
It doesn't change the current behavior, just make config more
explicit for large screen. A test depending on shell transition
is failed because its condition is stricter.

Moreover, because the size of 3-button bar is almost the same
as gestural mode which allows seamless rotation in large screen,
it should use the same config as gestural mode.

Bug: 249027315
Test: Enable shell transition, run
      atest SeamlessAppRotationTest
      on a large screen device.
Change-Id: I3fbf4d8c8fa477746abd18fd40ad65f55dc14d8f

Former-commit-id: 5b4279d2dc0cef3f5b7280161af742564c62d581