Override config should trigger intervention ignoring opt-in info
In the past we have an inconsistent behavior regarding override config
as it will ignore intervention setting such as `mAllowDownscale` but
still respect opt-in info as `mPerfModeOptedIn`. This will be confusing
and there is no way for OEM or game developers to test new interventions
for games without any pre-configured game mode device config.

Now they can instead first opt in the game modes temporarily if not
pre-configured (to make them available), then apply the override to
test. But they should reset the opt-in info and overrides after testing,
then communicate the interventions to OEMs.

This also fix the bug below where the override config used to contain
 full information including opt-in info that can be stale. Now it's
lightweight as it only contains GameModeConfiguration(s) and will be
 used together with default config in getConfig call.

Bug: b/253102835
Test: atest GameManagerServiceTests
Change-Id: Iee14d6eed07b16b6adb86b459edebdeca2b03fbc

Former-commit-id: 4f62571e7858c9a2b5dd8e5e342b00bcb45ea2bd